Update routes.tsx, app.tsx, and 3 more files...
