+++ frontend/config/routes.tsx
@@ -79,7 +79,8 @@ const defaultMenus = [
         ],
       },
       {
-        name: 'Layer',
+        key: 'WebMapServiceNestedLayer',
+        name: 'WebMapServiceNestedLayer',
         path: '/registry/WebMapService/:id/Layer',
         component: './registry/LayerTable',
         hideInMenu: true,
@@ -194,18 +195,3 @@ const defaultMenus = [
 ];
 
 export default defaultMenus;
-
-
-const traverseRoute = (route: MenuDataItem, name: string): MenuDataItem | undefined => {
-
-  if ( route.name === name ){
-    return route;
-  } else if ( route.routes && route.routes.length > 0 ){
-    return route.routes.find(_route => _route.name === name);
-  }
-    return undefined
-}
-
-export const findRouteByName = (name: string): MenuDataItem | undefined => {
-  return defaultMenus.find(route => traverseRoute(route, name));
-};
